gh-176: Splits time-sensitive tests out into a separate file (prior to removing/mocking them).
