+++ tests/unit/html/scheduler-timing-tests.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+        <title>Scheduler Timing Unit Tests</title>
+
+        <link rel="stylesheet" media="screen" href="../../../node_modules/infusion/tests/lib/qunit/css/qunit.css" />
+        <script src="../../../node_modules/infusion/tests/lib/qunit/js/qunit.js"></script>
+
+        <script src="../../../node_modules/jquery/dist/jquery.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/Fluid.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/FluidDebugging.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/FluidIoC.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/DataBinding.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/ModelTransformation.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/core/js/ModelTransformationTransforms.js"></script>
+        <script src="../../../node_modules/infusion/src/framework/enhancement/js/ContextAwareness.js"></script>
+
+        <script src="../../../src/core.js"></script>
+        <script src="../../../src/node-list.js"></script>
+        <script src="../../../src/evaluators.js"></script>
+        <script src="../../../src/parser.js"></script>
+        <script src="../../../src/scheduler.js"></script>
+        <script src="../../../src/web/webaudio-core.js"></script>
+        <script src="../../../src/web/audio-system.js"></script>
+        <script src="../../../src/web/buffer-writer.js"></script>
+        <script src="../../../src/web/input-device-manager.js"></script>
+        <script src="../../../src/web/native-node-manager.js"></script>
+        <script src="../../../src/web/output-manager.js"></script>
+        <script src="../../../src/ugens/core.js"></script>
+        <script src="../../../src/ugens/oscillators.js"></script>
+
+        <script src="../js/flocking-test-utils.js"></script>
+        <script src="../js/scheduler-timing-tests.js"></script>
+    </head>
+
+    <body id="body">
+        <h1 id="qunit-header">Scheduler Timing Unit Tests</h1>
+        <h2 id="qunit-banner"></h2>
+        <div id="qunit-testrunner-toolbar"></div>
+        <h2 id="qunit-userAgent"></h2>
+        <ol id="qunit-tests"></ol>
+
+        <!-- Test HTML -->
+        <div id="main" style="display: none;">
+
+        </div>
+    </body>
+</html>
