+++ tests/unit/js/scheduler-tests.js
@@ -56,167 +56,6 @@ var fluid = fluid || require("infusion"),
         }
     });
 
-
-    var checkScheduledCallback = function (expectedInterval, reportedInterval, scheduledAt, receivedAt, maxOutlier) {
-        var duration = receivedAt - scheduledAt,
-            minDur = expectedInterval - maxOutlier,
-            maxDur = expectedInterval + maxOutlier;
-
-        equal(reportedInterval, expectedInterval,
-            "The callback should return the correct scheduled time.");
-        ok(duration >= minDur && duration <= maxDur,
-            "The callback should be fired at the scheduled time, within a tolerance of " +
-            maxOutlier + "ms. Actual interval was: " + (duration - expectedInterval) + "ms.");
-
-        return Math.abs(duration - expectedInterval);
-    };
-
-    var testOnceSchedule = function (testSpec) {
-        var numRunsComplete = 0,
-            schedulerType = testSpec.scheduler.type,
-            schedulerOpts = testSpec.scheduler.options,
-            scheduler = fluid.invokeGlobalFunction(schedulerType, [schedulerOpts]),
-            actualAverageDrift = 0,
-            scheduledAt,
-            scheduleFn,
-            scheduledAction;
-
-        scheduleFn = function () {
-            scheduler[testSpec.interval](testSpec.time, scheduledAction);
-            return Date.now();
-        };
-
-        scheduledAction = function (scheduledTime) {
-            numRunsComplete++;
-            actualAverageDrift += checkScheduledCallback(
-                testSpec.time, scheduledTime, scheduledAt, Date.now(), testSpec.maxOutlier);
-            if (numRunsComplete < testSpec.runs) {
-                scheduledAt = scheduleFn();
-            } else {
-                actualAverageDrift = actualAverageDrift / numRunsComplete;
-                equal(numRunsComplete, testSpec.runs,
-                    "The scheduled callback should be invoked only once.");
-                ok(actualAverageDrift <= testSpec.averageDrift,
-                    "The average drift of the scheduler should be no more than " +
-                    testSpec.averageDrift + "ms. Actual was: " + actualAverageDrift + "ms.");
-                start();
-            }
-        };
-
-        scheduledAt = scheduleFn();
-
-        return scheduler;
-    };
-
-    asyncTest("Scheduling raw functions with once().", function () {
-        expect(22);
-
-        sked = testOnceSchedule({
-            runs: 10,
-            interval: "once",
-            time: 100,
-            maxOutlier: 65,
-            averageDrift: 10,
-            scheduler: {
-                type: "flock.scheduler.async",
-                options: {
-                    components: {
-                        timeConverter: {
-                            type: "flock.convert.ms"
-                        }
-                    }
-                }
-            }
-        });
-    });
-
-    asyncTest("flock.scheduler.once() multiple listeners, different intervals", function () {
-        // TODO: Cut and pastage and inconsistencies everywhere!
-        var scheduledDelays = [100, 200],
-            tolerance = 50, // TODO: Insanely high.
-            fired = {},
-            makeRecordingListener,
-            assertWithinTolerance,
-            testingListenerImpl,
-            listener1,
-            listener2,
-            testingListener,
-            scheduledAt;
-
-        sked = flock.scheduler.async({
-            components: {
-                timeConverter: {
-                    type: "flock.convert.ms"
-                }
-            }
-        });
-
-        makeRecordingListener = function (record, prop) {
-            return function () {
-                record[prop] = Date.now() - scheduledAt;
-            };
-        };
-
-        assertWithinTolerance = function (actual, scheduled, tolerance, msgPrefix) {
-            ok(actual >= scheduled - tolerance &&
-                actual <= scheduled + tolerance,
-                msgPrefix + " should be scheduled at the expected time, within a tolerance of " +
-                tolerance + "ms." + " Actual: " + (actual - scheduled) + "ms.");
-        };
-
-        testingListenerImpl = function () {
-            assertWithinTolerance(fired.listener1, scheduledDelays[0], tolerance, "The first callback");
-            assertWithinTolerance(fired.listener2, scheduledDelays[1], tolerance, "The second callback");
-            start();
-        };
-
-        listener1 = sked.once(100, makeRecordingListener(fired, "listener1"));
-        listener2 = sked.once(200, makeRecordingListener(fired, "listener2"));
-        testingListener = sked.once(300, testingListenerImpl);
-        scheduledAt = Date.now();
-    });
-
-    asyncTest("flock.scheduler.async.repeat()", function () {
-        var expectedInterval = 100,
-            numRuns = 100,
-            runs = 0,
-            lastFired = 0,
-            mistimingTolerance = 60, // TODO: This value is excessively high.
-            callback;
-
-        expect(2 * numRuns);
-
-        sked = flock.scheduler.async({
-            components: {
-                timeConverter: {
-                    type: "flock.convert.ms"
-                }
-            }
-        });
-
-        callback = function () {
-            var now = Date.now(),
-                actualInterval = now - lastFired;
-
-            if (runs >= numRuns) {
-                sked.repeatScheduler.clearInterval(expectedInterval);
-                start();
-                return;
-            }
-
-            ok(actualInterval >= expectedInterval - mistimingTolerance,
-                "The scheduled callback should be fired no earlier than " + mistimingTolerance + " ms.");
-            ok(actualInterval <= expectedInterval + mistimingTolerance,
-                "The scheduled callback should be fired no later than " + mistimingTolerance + " ms.");
-
-            runs++;
-            lastFired = Date.now();
-        };
-
-        sked.repeat(expectedInterval, callback);
-        lastFired = Date.now();
-    });
-
     asyncTest("flock.scheduler.async.repeat() multiple listeners", function () {
         var interval = 100,
             numRuns = 10,
